* master: (23 commits)
  Adds indexes to database (#525)
  Swap out floats (#527)
  Improved chart formatting (#526)
  Make transaction expiration a config (#520)
  Don't retry succeeded tasks (#523)
  Bump cryptography from 3.0 to 3.2 in /eth_worker (#522)
  Negative balance disbursements display (#524)
  Fix user deletion pending txn bug (#519)
  Add unstarted status (#503)
  commenting out last seen (#516)
  export phone in xlsx (#515)
  Improving analytics load (#513)
  Export bugfix (#510)
  Fix transfer graph currency bug (#508)
  Improve export functionality (#454)
  Remove redundant boolean/list representations of permissions from account creation API (#501)
  searchable filter/group by modules (#506)
  configurable option on data wrapper (#504)
  zero-pad dates til today (#502)
  Accelerate any call relying on transfer usage (#497)
  ...

# Conflicts:
#	app/package.json
#	app/server/models/user.py
#	config.py

+++ app/server/api/credit_transfer_api.py
@@ -2,6 +2,7 @@ from flask import Blueprint, request, make_response, jsonify, g
 from flask.views import MethodView
 from sqlalchemy import or_, not_
 import json
+from decimal import Decimal
 from uuid import uuid4
 from server import db
 from server.models.token import Token
@@ -184,7 +185,7 @@ class CreditTransferAPI(MethodView):
         queue = 'low-priority'
 
         transfer_type = post_data.get('transfer_type')
-        transfer_amount = abs(round(float(post_data.get('transfer_amount') or 0),6))
+        transfer_amount = abs(round(Decimal(post_data.get('transfer_amount') or 0),6))
         token_id = post_data.get('token_id')
         target_balance = post_data.get('target_balance')
