+++ app/server/models/organisation.py
@@ -4,6 +4,7 @@ from sqlalchemy.dialects.postgresql import ARRAY
 from sqlalchemy import type_coerce
 import pendulum
 import secrets
+from decimal import Decimal
 
 from server import db, bt
 from server.models.utils import ModelBase, organisation_association_table
@@ -97,7 +98,7 @@ class Organisation(ModelBase):
 
     @property
     def default_disbursement(self):
-        return float((self._default_disbursement_wei or 0) / int(1e16))
+        return Decimal((self._default_disbursement_wei or 0) / int(1e16))
 
     @default_disbursement.setter
     def default_disbursement(self, val):
